display the real ap cost or X for none
