+++ src/game/ui/game_interface.rs
@@ -1,8 +1,8 @@
 use bevy::prelude::*;
 
 use crate::{
-    game::{pieces::components::{Health, Monster}, player::{Player, Cursor}, combat::components::ActionPoints},
-    globals::{INTERFACE_GLOBAL_PLAYER_NAME_FONT_SIZE, TILE_WIDTH_HALF, TILE_HEIGHT_HALF, CHAR_SIZE}, render::components::GameCursorRender
+    game::{pieces::components::{Health, Monster, Occupier}, player::{Player, Cursor}, combat::{components::ActionPoints, event_systems::get_ap_cost}, tileboard::components::BoardPosition},
+    globals::{INTERFACE_GLOBAL_PLAYER_NAME_FONT_SIZE, TILE_WIDTH_HALF, TILE_HEIGHT_HALF, CHAR_SIZE}, render::components::GameCursorRender, map_builders::map::Map
 };
 
 use super::components::{InterfaceGame, UiEnemyHp, UiActionPointsOnCursor};
@@ -11,8 +11,6 @@ const INTERFACE_HP_CHUNK_HEIGHT: f32 = 16.;
 const INTERFACE_HP_CHUNK_WIDTH: f32 = 8.;
 
 const INTERFACE_HP_CHUNK_MAX: u32 = 20;
-
-
 
 
 
@@ -49,10 +47,26 @@ pub fn display_action_points_on_cursor(
     asset_server: Res<AssetServer>,
     cursor: Res<Cursor>,
     camera_q: Query<(&Camera, &GlobalTransform)>, 
-    query_game_cursor: Query<(&GameCursorRender, &mut Transform)>,
+    query_game_cursor: Query<&mut Transform, With<GameCursorRender>>,
     interface_query: Query<Entity, With<UiActionPointsOnCursor>>,
+    player_q: Query<Entity, With<Player>>,
+    mut query_character: Query<(&ActionPoints, &BoardPosition)>,
+    query_occupied: Query<&BoardPosition, With<Occupier>>,
+    board: Res<Map>,
+    //tile_position: Vector2Int,
+    //entity: Entity,
 ){
     clear_action_points_cursor_ui(&mut commands, interface_query);
+
+    let Ok(player) = player_q.get_single() else { return };
+    let ap_cost_result = get_ap_cost(query_character, query_occupied, board, cursor.grid_position, player);
+ 
+    let mut ap_result = 'x';
+    if let Some(ap_cost) = ap_cost_result.cost {
+        let Some(ap_char) = char::from_digit(ap_cost, 10) else { return };
+        println!("Ap char is {:?}", ap_char);
+        ap_result = ap_char;
+    }
 
     let (camera, camera_transform) = camera_q.single();
     let Some(screen_size) = camera.logical_viewport_size() else { return };    // What we can see in the screen. Some(Vec2(1422.0, 800.0) So 0,1422 and 1422, 800.0 for each corner.
@@ -60,28 +74,11 @@ pub fn display_action_points_on_cursor(
 
     //println!("Camera physical viewport size is {:?}", screen_size);
 
-    for (cursor, transform) in query_game_cursor.iter() {
+    for transform in query_game_cursor.iter() {
         let Some(screen_position) = camera.world_to_viewport(camera_transform, transform.translation)  else { continue };
         //If not in screen, we don't display.
         if screen_position.x < 0.0 || screen_position.x > screen_size.x || screen_position.y < 0.0 || screen_position.y > screen_size.y { continue};
-      
-      /*
-        let left =screen_position.x - (TILE_WIDTH_HALF as f32);
-        //let right =screen_size.x - screen_position.x;
-        let top =screen_position.y - (TILE_HEIGHT_HALF as f32); // REMEMBER : world = y goes from bottom to top (++)
-        //let bottom = screen_size.y - screen_position.y;
-        let width = (health.max as f32 * INTERFACE_HP_CHUNK_WIDTH) / 2.0; //INTERFACE_HP_CHUNK_WIDTH * (health.max as f32) / 2.0;
-        let height = INTERFACE_HP_CHUNK_HEIGHT/ 2.0;
-        //println!("Character screen position is : {:?}", screen_position);
-        //println!("left : {:?}, right : {:?}, top : {:?}, bottom : {:?}, width: {:?}, height: {:?}", left ,right ,top ,bottom, width, height );
-
-
-        let grow = (health.max as f32 * INTERFACE_HP_CHUNK_WIDTH) / 2.0;
-     */    
-    
-        //let (camera, camera_transform) = camera_q.single();
-        //let Some(screen_size) = camera.logical_viewport_size() else { return };    // What we can see in the screen. Some(Vec2(1422.0, 800.0) So 0,1422 and 1422, 800.0 for each corner.
-        
+  
         let left = screen_position.x + (CHAR_SIZE as f32 / 2.0);
         let top = screen_position.y + (CHAR_SIZE as f32 / 2.0); 
 
@@ -111,7 +108,7 @@ pub fn display_action_points_on_cursor(
 
         let cursor_action_display = commands.spawn(
             TextBundle::from_section(
-                format!("5"),     //("{}",action_points),
+                format!("{}", ap_result),     //("{}",action_points),
                 TextStyle { 
                     font: asset_server.load("fonts/PressStart2P-vaV7.ttf"),
                     font_size: INTERFACE_GLOBAL_PLAYER_NAME_FONT_SIZE,
